+++ lib/ts/recipe/thirdparty/types.ts
@@ -176,6 +176,7 @@ export interface APIInterface {
               provider: TypeProvider;
               code: string;
               redirectURI: string;
+              authCodeResponse?: any;
               options: APIOptions;
           }) => Promise<
               | {