+++ lib/ts/recipe/thirdpartyemailpassword/recipeImplementation/index.ts
@@ -4,29 +4,32 @@ import EmailPasswordImplemenation from "../../emailpassword/recipeImplementation
 import ThirdPartyImplemenation from "../../thirdparty/recipeImplementation";
 import { RecipeInterface as ThirdPartyRecipeInterface } from "../../thirdparty";
 import { Querier } from "../../../querier";
+import DerivedEP from "./emailPasswordRecipeImplementation";
+import DerivedTP from "./thirdPartyRecipeImplementation";
 
 export default function getRecipeInterface(
     emailPasswordQuerier: Querier,
     thirdPartyQuerier?: Querier
 ): RecipeInterface {
-    let emailPasswordImplementation = EmailPasswordImplemenation(emailPasswordQuerier);
-    let thirdPartyImplementation: undefined | ThirdPartyRecipeInterface;
+    let originalEmailPasswordImplementation = EmailPasswordImplemenation(emailPasswordQuerier);
+    let originalThirdPartyImplementation: undefined | ThirdPartyRecipeInterface;
     if (thirdPartyQuerier !== undefined) {
-        thirdPartyImplementation = ThirdPartyImplemenation(thirdPartyQuerier);
+        originalThirdPartyImplementation = ThirdPartyImplemenation(thirdPartyQuerier);
     }
+
     return {
         signUp: async function (input: {
             email: string;
             password: string;
         }): Promise<{ status: "OK"; user: User } | { status: "EMAIL_ALREADY_EXISTS_ERROR" }> {
-            return await emailPasswordImplementation.signUp(input);
+            return await originalEmailPasswordImplementation.signUp.bind(DerivedEP(this))(input);
         },
 
         signIn: async function (input: {
             email: string;
             password: string;
         }): Promise<{ status: "OK"; user: User } | { status: "WRONG_CREDENTIALS_ERROR" }> {
-            return emailPasswordImplementation.signIn(input);
+            return originalEmailPasswordImplementation.signIn.bind(DerivedEP(this))(input);
         },
 
         signInUp: async function (input: {
@@ -43,30 +46,36 @@ export default function getRecipeInterface(
                   error: string;
               }
         > {
-            if (thirdPartyImplementation === undefined) {
+            if (originalThirdPartyImplementation === undefined) {
                 throw new Error("No thirdparty provider configured");
             }
-            return thirdPartyImplementation.signInUp(input);
+            return originalThirdPartyImplementation.signInUp.bind(DerivedTP(this))(input);
         },
 
         getUserById: async function (input: { userId: string }): Promise<User | undefined> {
-            let user: User | undefined = await emailPasswordImplementation.getUserById(input);
+            let user: User | undefined = await originalEmailPasswordImplementation.getUserById.bind(DerivedEP(this))(
+                input
+            );
             if (user !== undefined) {
                 return user;
             }
-            if (thirdPartyImplementation === undefined) {
+            if (originalThirdPartyImplementation === undefined) {
                 return undefined;
             }
-            return await thirdPartyImplementation.getUserById(input);
+            return await originalThirdPartyImplementation.getUserById.bind(DerivedTP(this))(input);
         },
 
         getUsersByEmail: async function ({ email }: { email: string }): Promise<User[]> {
-            let userFromEmailPass: User | undefined = await emailPasswordImplementation.getUserByEmail({ email });
+            let userFromEmailPass: User | undefined = await originalEmailPasswordImplementation.getUserByEmail.bind(
+                DerivedEP(this)
+            )({ email });
 
-            if (thirdPartyImplementation === undefined) {
+            if (originalThirdPartyImplementation === undefined) {
                 return userFromEmailPass === undefined ? [] : [userFromEmailPass];
             }
-            let usersFromThirdParty: User[] = await thirdPartyImplementation.getUsersByEmail({ email });
+            let usersFromThirdParty: User[] = await originalThirdPartyImplementation.getUsersByEmail.bind(
+                DerivedTP(this)
+            )({ email });
 
             if (userFromEmailPass !== undefined) {
                 return [...usersFromThirdParty, userFromEmailPass];
@@ -78,20 +87,20 @@ export default function getRecipeInterface(
             thirdPartyId: string;
             thirdPartyUserId: string;
         }): Promise<User | undefined> {
-            if (thirdPartyImplementation === undefined) {
+            if (originalThirdPartyImplementation === undefined) {
                 return undefined;
             }
-            return thirdPartyImplementation.getUserByThirdPartyInfo(input);
+            return originalThirdPartyImplementation.getUserByThirdPartyInfo.bind(DerivedTP(this))(input);
         },
 
         createResetPasswordToken: async function (input: {
             userId: string;
         }): Promise<{ status: "OK"; token: string } | { status: "UNKNOWN_USER_ID_ERROR" }> {
-            return emailPasswordImplementation.createResetPasswordToken(input);
+            return originalEmailPasswordImplementation.createResetPasswordToken.bind(DerivedEP(this))(input);
         },
 
         resetPasswordUsingToken: async function (input: { token: string; newPassword: string }) {
-            return emailPasswordImplementation.resetPasswordUsingToken(input);
+            return originalEmailPasswordImplementation.resetPasswordUsingToken.bind(DerivedEP(this))(input);
         },
 
         updateEmailOrPassword: async function (input: {
@@ -99,7 +108,7 @@ export default function getRecipeInterface(
             email?: string;
             password?: string;
         }): Promise<{ status: "OK" | "UNKNOWN_USER_ID_ERROR" | "EMAIL_ALREADY_EXISTS_ERROR" }> {
-            return emailPasswordImplementation.updateEmailOrPassword(input);
+            return originalEmailPasswordImplementation.updateEmailOrPassword.bind(DerivedEP(this))(input);
         },
     };
 }